feat: implemented moving from u64 to u128 (#323)
* implemented moving from u64 to u128

* remove generated file

* fix parsing in CurrencyCollection

* change type of parameter from_lt

* fix grams

* fix attached amount

* remove u128 from_lt

* switch to another branch in nekoton

* update version of nekoton_bridge

* try to fix ci git diff

---------

+++ packages/nekoton_bridge/rust/src/nekoton_wrapper/core/jetton_wallet/mod.rs
@@ -190,7 +190,7 @@ impl JettonWalletBoxTrait for JettonWalletBox {
         attached_amount: Option<String>,
     ) -> anyhow::Result<String> {
         let attached_amount = attached_amount.unwrap_or("100000000".to_string()); // DEFAULT if 0.1 TON
-        let attached_amount = u64::from_str(&attached_amount).handle_error()?;
+        let attached_amount = u128::from_str(&attached_amount).handle_error()?;
 
         let destination = parse_address(destination)?;
         let remaining_gas_to = parse_address(remaining_gas_to)?;
